# packages/types/openapi.yaml
openapi: '3.1.0'
info:
  title: Self-Hosted PaaS API
  version: '1.0.0'
  description: API for managing projects, services, databases, and deployments.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/projects/{projectId}/services:
    get:
      summary: List services
      tags: [Services]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: List of services
    post:
      summary: Create service
      tags: [Services]
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                type: { type: string, enum: [git, image] }
                source_repo: { type: string }
                source_image: { type: string }
                build_type: { type: string, enum: [dockerfile, buildpacks, railpack, image] }
                builder_image: { type: string }
                custom_buildpacks: { type: array, items: { type: string } }
              required: [name, type]
      responses:
        '201':
          description: Service created
...